#107 new in-mem ds plugged in to auth server script
